# artifacts/api-server/src/api-zod/generated/types/createConfigRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrepMap Exam Roadmap Platform API
 * OpenAPI spec version: 0.3.0
 */
import type { CreateConfigRequestExam } from "./createConfigRequestExam";

export interface CreateConfigRequest {
  universityId: string;
  year: string;
  branch: string;
  subject: string;
  exam: CreateConfigRequestExam;
}