# artifacts/exam-roadmap/src/api-client/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrepMap Exam Roadmap Platform API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginRequest {
  collegeId: string;
  password: string;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  admin: "admin",
  student: "student",
  super_student: "super_student",
} as const;

export interface LoginResponse {
  id: string;
  universityId: string;
  branch: string;
  year: string;
  role: LoginResponseRole;
  mustResetPassword?: boolean;
  securityQuestionSet?: boolean;
  onboardingRequired?: boolean;
}

export interface ResetPasswordRequest {
  collegeId: string;
  currentPassword: string;
  newPassword: string;
}

export type CreateConfigRequestExam =
  (typeof CreateConfigRequestExam)[keyof typeof CreateConfigRequestExam];

export const CreateConfigRequestExam = {
  mid1: "mid1",
  mid2: "mid2",
  endsem: "endsem",
} as const;

export interface CreateConfigRequest {
  universityId: string;
  year: string;
  branch: string;
  subject: string;
  exam: CreateConfigRequestExam;
}

export interface UploadConfigFilesRequest {
  syllabusFileUrl: string;
  paperFileUrls: string[];
}

export type ConfigExam = (typeof ConfigExam)[keyof typeof ConfigExam];

export const ConfigExam = {
  mid1: "mid1",
  mid2: "mid2",
  endsem: "endsem",
} as const;

export type ConfigStatus = (typeof ConfigStatus)[keyof typeof ConfigStatus];

export const ConfigStatus = {
  draft: "draft",
  live: "live",
} as const;

export interface Config {
  id: string;
  universityId: string;
  year: string;
  branch: string;
  subject: string;
  exam: ConfigExam;
  status: ConfigStatus;
  createdBy: string;
  createdAt?: string;
  syllabusFileUrl?: string | null;
  paperFileUrls?: string[] | null;
}

export type GenerationStatusStatus =
  (typeof GenerationStatusStatus)[keyof typeof GenerationStatusStatus];

export const GenerationStatusStatus = {
  idle: "idle",
  parsing: "parsing",
  generating: "generating",
  complete: "complete",
  error: "error",
} as const;

export interface GenerationStatus {
  configId: string;
  status: GenerationStatusStatus;
  progress: number;
  total: number;
  currentStep: string;
  error?: string | null;
}

export type NodeType = (typeof NodeType)[keyof typeof NodeType];

export const NodeType = {
  unit: "unit",
  topic: "topic",
  subtopic: "subtopic",
} as const;

export interface Node {
  id: string;
  configId: string;
  title: string;
  type: NodeType;
  parentId?: string | null;
  explanation?: string | null;
  sortOrder?: string;
}

export type SubtopicQuestionMarkType =
  (typeof SubtopicQuestionMarkType)[keyof typeof SubtopicQuestionMarkType];

export const SubtopicQuestionMarkType = {
  Foundational: "Foundational",
  Applied: "Applied",
} as const;

export interface SubtopicQuestion {
  id: number;
  markType: SubtopicQuestionMarkType;
  question: string;
  answer: string;
  isStarred?: boolean;
  starSource?: "none" | "auto" | "manual";
}

export interface SubtopicContent {
  id: string;
  nodeId: string;
  explanation: string;
  questions: SubtopicQuestion[];
}

export type UpdateQuestionMarkType =
  (typeof UpdateQuestionMarkType)[keyof typeof UpdateQuestionMarkType];

export const UpdateQuestionMarkType = {
  Foundational: "Foundational",
  Applied: "Applied",
} as const;

export interface UpdateQuestion {
  id?: number | null;
  markType: UpdateQuestionMarkType;
  question: string;
  answer: string;
  isStarred?: boolean;
  starSource?: "none" | "auto" | "manual";
}

export interface UpdateSubtopicRequest {
  explanation: string;
  questions: UpdateQuestion[];
}

export interface EventPayload {
  userId: string;
  universityId: string;
  year: string;
  branch: string;
  exam: string;
  configId: string;
  topicId: string;
  subtopicId: string;
}

export interface SubtopicStat {
  subtopicId: string;
  subtopicTitle: string;
  eventCount: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type GetConfigsParams = {
  universityId?: string;
  status?: GetConfigsStatus;
};

export type GetConfigsStatus =
  (typeof GetConfigsStatus)[keyof typeof GetConfigsStatus];

export const GetConfigsStatus = {
  draft: "draft",
  live: "live",
} as const;

export type GetNodesParams = {
  configId: string;
};