# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: PrepMap Exam Roadmap Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /configs:
    get:
      operationId: getConfigs
      tags: [configs]
      summary: Get exam configurations
      parameters:
        - name: universityId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, live]
      responses:
        "200":
          description: List of configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Config"
    post:
      operationId: createConfig
      tags: [configs]
      summary: Create a new config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConfigRequest"
      responses:
        "201":
          description: Config created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
components:
  schemas:
...
    CreateConfigRequest:
      type: object
      properties:
        universityId:
          type: string
        year:
          type: string
        branch:
          type: string
        subject:
          type: string
        exam:
          type: string
          enum: [mid1, mid2, endsem]
      required:
        - universityId
        - year
        - branch
        - subject
        - exam
...
    Config:
      type: object
      properties:
        id:
          type: string
        universityId:
          type: string
        year:
          type: string
        branch:
          type: string
        subject:
          type: string
        exam:
          type: string
          enum: [mid1, mid2, endsem]
        status:
          type: string
          enum: [draft, live]
        createdBy:
          type: string
        createdAt:
          type: string
        syllabusFileUrl:
          type: string
          nullable: true
        paperFileUrls:
          type: array
          items:
            type: string
          nullable: true
      required:
        - id
        - universityId
        - year
        - branch
        - subject
        - exam
        - status
        - createdBy
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"